Models work
Re-adding models folder creation, migrates, game key (unfinished)
