+++ catan/models/__init__.py
@@ -1,0 +1,4 @@
+from .game import Game
+from .board import Board
+from .corner import Corner
+from .tile import Tile
