+++ catan/models/game.py
@@ -1,0 +1,9 @@
+from django.db import models
+"""
+Game overhead. It stores the players, along with flags for turn progression
+and multi-step interaction.
+"""
+class Game(models.Model):
+    game_key = models.UUIDField(primary_key=True)
+    number_players = models.IntegerField()
+    turn = models.IntegerField()
